+++ recording/CHANGELOG.md
@@ -2,3 +2,4 @@
 
 ## 0.1.0-dev
 * Initial release [#356](https://github.com/jellyfish-dev/membrane_rtc_engine/pull/356)
+* Add crash groups, fix adding nonexistent tracks to state, fix READMEs [#366](https://github.com/jellyfish-dev/membrane_rtc_engine/pull/366)